Added Middleware, Controllers and Fix Users Model and Router
