+++ server/models/user.js
@@ -2,6 +2,11 @@ const mongoose = require('mongoose');
 
 const userSchema = mongoose.Schema({
   _id: mongoose.Schema.Types.ObjectId,
+  name: {
+    type: String, 
+    required: true, 
+    unique: true
+  },
   email: { 
     type: String, 
     required: true,
